% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{somaticflags}
\alias{somaticflags}
\title{Somatic Flags Dataset}
\format{
A vector with 139 gene names:
}
\source{
\doi{10.1038/nature12213} and \doi{10.1186/s12920-014-0064-y}
}
\usage{
somaticflags
}
\description{
a list of genes which are frequently mutated in somatic cancer datasets but are unlikely to drive disease.
}
\examples{
data(somaticflags)
head(somaticflags)
}
\keyword{datasets}
