% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{somaticflags_df}
\alias{somaticflags_df}
\title{Somatic Flags Dataset}
\format{
A data.frame with 139 gene names:

\describe{
\item{Gene}{Genes frequently mutated in somatic cancer datasets but that are unlikely to drive disease}
\item{Source}{Supporting papers for the gene being a somatic flag}
\item{Reason}{Explanation for the gene being considered a somatic flag}
...
}
}
\source{
\url{https://doi.org/10.1038/nature12213} and \url{https://doi.org/10.1186/s12920-014-0064-y}
}
\usage{
somaticflags_df
}
\description{
a list of genes which are frequently mutated in somatic cancer datasets but are unlikely to drive disease.
}
\examples{
data(somaticflags_df)
head(somaticflags_df)
}
\keyword{datasets}
